Restructuring subscribers
